newsapi kotlin: reorganize packages
