+++ DatabindingWithNewsApiKotlin/app/src/main/java/com/enpassio/databindingwithnewsapikotlin/data/NewsRepository.kt
@@ -4,6 +4,7 @@ import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.MutableLiveData
 import android.os.AsyncTask
 import android.util.Log
+import com.enpassio.databindingwithnewsapikotlin.model.Article
 import com.enpassio.databindingwithnewsapikotlin.utils.fetchArticles
