+++ DatabindingWithNewsApiKotlin/app/src/main/java/com/enpassio/databindingwithnewsapikotlin/ui/ArticleDetailsFragment.kt
@@ -14,7 +14,6 @@ import android.view.View
 import android.view.ViewGroup
 import com.enpassio.databindingwithnewsapikotlin.R
 import com.enpassio.databindingwithnewsapikotlin.databinding.FragmentDetailsBinding
-import com.enpassio.databindingwithnewsapikotlin.viewmodel.MainViewModel
 
 
 class ArticleDetailsFragment : Fragment() {
